edit tao hoa don ko dc san pham trong

+++ ASM_NET105_BanTui/Controllers/GioHangController.cs
@@ -47,7 +47,6 @@ namespace ASM_NET105_BanTui.Controllers
             var loginData = HttpContext.Session.GetString("UserId");
             var userCart = context.GioHangCT.Where(temp => temp.ID_User == Guid.Parse(loginData)).ToList();
 
-            
             foreach (var item in userCart)
             {
                 var product = context.SanPham.FirstOrDefault(temp => temp.ID_SanPham == item.ID_SanPham);
@@ -58,35 +57,43 @@ namespace ASM_NET105_BanTui.Controllers
                 }
             }
 
-            var hoaDon = new HoaDon()
+            if(userCart.Count > 0)
             {
-                ID_HoaDon = Guid.NewGuid(),
-                ID_User = Guid.Parse(loginData),
-                TongTien = userCart.Sum(temp => temp.SoLuong * temp.SanPham.GiaNiemYet),
-                TrangThai = StatusOfBillOptions.Paid.ToString(),
-                NgayThanhToan = DateTime.Now
-            };
-
-            HttpContext.Session.SetString("hoaDonId", hoaDon.ID_HoaDon.ToString());
-
-            foreach (var item in userCart)
-            {
-                var hoaDonCT = new HoaDonCT()
+                var hoaDon = new HoaDon()
                 {
-                    ID_HoaDonCT = Guid.NewGuid(),
-                    ID_HoaDon = hoaDon.ID_HoaDon,
-                    ID_SanPham = item.ID_SanPham,
-                    GiaBan = item.SanPham.GiaNiemYet,
-                    SoLuong = item.SoLuong,
+                    ID_HoaDon = Guid.NewGuid(),
+                    ID_User = Guid.Parse(loginData),
+                    TongTien = userCart.Sum(temp => temp.SoLuong * temp.SanPham.GiaNiemYet),
+                    TrangThai = StatusOfBillOptions.Paid.ToString(),
+                    NgayThanhToan = DateTime.Now
                 };
-                context.HoaDonCT.Add(hoaDonCT);
-            }
-            
-            context.HoaDon.Add(hoaDon);
-            context.GioHangCT.RemoveRange(userCart);
-            context.SaveChanges();
 
-            return RedirectToAction("IndexKH", "SanPham");
+                HttpContext.Session.SetString("hoaDonId", hoaDon.ID_HoaDon.ToString());
+
+                foreach (var item in userCart)
+                {
+                    var hoaDonCT = new HoaDonCT()
+                    {
+                        ID_HoaDonCT = Guid.NewGuid(),
+                        ID_HoaDon = hoaDon.ID_HoaDon,
+                        ID_SanPham = item.ID_SanPham,
+                        GiaBan = item.SanPham.GiaNiemYet,
+                        SoLuong = item.SoLuong,
+                    };
+                    context.HoaDonCT.Add(hoaDonCT);
+                }
+
+                context.HoaDon.Add(hoaDon);
+                context.GioHangCT.RemoveRange(userCart);
+                context.SaveChanges();
+
+                return RedirectToAction("Index", "GioHang");
+            }
+            else
+            {
+                TempData["Message2"] = "Bạn chưa có sản phẩm nào trong giỏ hàng !";
+            }
+            return RedirectToAction("Index", "GioHang");
         }
     }
 }
